break up UploadQueue into separate components
